[C] Separate contributor detail layouts into base, item, and collection
