+++ packages/frontend/components/composed/contributor/ContributorDetail/ContributorDetail.styles.tsx
