+++ packages/frontend/components/composed/contributor/ContributorDetail/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ContributorDetail";
