+++ packages/frontend/components/composed/contributor/ContributorDetailLayout/ContributorDetailLayout.tsx
@@ -1,132 +1,33 @@
-import React, { useMemo } from "react";
+import React from "react";
 import { graphql } from "react-relay";
-import ReactMarkdown from "react-markdown";
-import { useTranslation } from "react-i18next";
 import { useMaybeFragment } from "@wdp/lib/api/hooks";
-import get from "lodash/get";
-import ContributorName from "../ContributorName";
-import ContributorAvatar from "../ContributorAvatar";
-import * as Styled from "./ContributorDetailLayout.styles";
-import ContributionItem from "./ContributionItem/ContributionItem";
-import { ExternalLink } from "components/atomic";
+import ContributorDetail from "components/composed/contributor/ContributorDetail";
+import ContributorHTMLHead from "components/composed/contributor/ContributorHTMLHead";
 import { ContributorDetailLayoutFragment$key } from "@/relay/ContributorDetailLayoutFragment.graphql";
-import { BackToTopBlock } from "components/layout";
+import BaseLayout from "components/composed/base/BaseLayout";
 
 export default function ContributorDetailLayout({ data }: Props) {
-  const contributor = useMaybeFragment(fragment, data);
-  const { t } = useTranslation();
+  const layoutData = useMaybeFragment(fragment, data);
 
-  const contributionCount = useMemo(
-    () =>
-      get(contributor, "collectionContributionCount", 0) +
-      get(contributor, "itemContributionCount", 0),
-    [contributor]
-  );
-
-  return contributor ? (
-    <>
-      <section className="a-bg-custom10">
-        <Styled.ContributorWrapper className="l-container-wide">
-          <Styled.NameHeader as="header">
-            {contributor.image?.storage && (
-              <ContributorAvatar data={contributor.image} />
-            )}
-            <h3>
-              <ContributorName data={contributor} />
-            </h3>
-          </Styled.NameHeader>
-          {contributor.title && (
-            <p className="t-label-lg">{contributor.title}</p>
-          )}
-          {contributor.affiliation && (
-            <p className="t-copy-lighter">{contributor.affiliation}</p>
-          )}
-          {contributor.bio && (
-            <div className="t-copy-light">
-              <ReactMarkdown>{contributor.bio}</ReactMarkdown>
-            </div>
-          )}
-          {contributor.links && contributor.links.length > 0 && (
-            <ul className="t-label-sm">
-              {contributor.links.map(({ title, url }, i) => (
-                <li key={i}>
-                  <ExternalLink href={url}>{title}</ExternalLink>
-                </li>
-              ))}
-            </ul>
-          )}
-        </Styled.ContributorWrapper>
-      </section>
-      {contributionCount > 0 && (
-        <section className="l-container-wide">
-          <Styled.ContributionsWrapper
-            as={contributionCount > 6 ? BackToTopBlock : "div"}
-          >
-            <Styled.ContributionsHeader>
-              {t("layouts.contributions_header")}
-            </Styled.ContributionsHeader>
-            <Styled.ContributionList>
-              {contributor.collectionContributions.nodes.map(
-                ({ collection }, i) => (
-                  <Styled.ContributionItem key={i}>
-                    <ContributionItem data={collection} />
-                  </Styled.ContributionItem>
-                )
-              )}
-              {contributor.itemContributions.nodes.map(({ item }, i) => (
-                <Styled.ContributionItem key={i}>
-                  <ContributionItem data={item} />
-                </Styled.ContributionItem>
-              ))}
-            </Styled.ContributionList>
-          </Styled.ContributionsWrapper>
-        </section>
-      )}
-    </>
+  return layoutData ? (
+    <BaseLayout data={layoutData}>
+      <ContributorHTMLHead data={layoutData.contributor} />
+      <ContributorDetail data={layoutData.contributor} />
+    </BaseLayout>
   ) : null;
 }
 
 interface Props {
   data?: ContributorDetailLayoutFragment$key | null;
-  backRoute?: string;
-  backLabel?: string;
 }
 
 const fragment = graphql`
-  fragment ContributorDetailLayoutFragment on Contributor {
-    ...ContributorNameFragment
-    ... on Contributor {
-      bio
-      collectionContributionCount
-      collectionContributions {
-        nodes {
-          role
-          collection {
-            ...ContributionItemFragment
-          }
-        }
-      }
-      itemContributionCount
-      itemContributions {
-        nodes {
-          role
-          item {
-            ...ContributionItemFragment
-          }
-        }
-      }
-      image {
-        storage
-        ...ContributorAvatarFragment
-      }
-      links {
-        title
-        url
-      }
+  fragment ContributorDetailLayoutFragment on Query {
+    contributor(slug: $slug) {
+      ...ContributorHTMLHeadFragment
+      ...ContributorDetailFragment
     }
-    ... on PersonContributor {
-      affiliation
-      title
-    }
+
+    ...BaseLayoutFragment
   }
 `;
