+++ packages/frontend/components/composed/contributor/ContributorDetailLayout/index.ts
@@ -1,1 +1,3 @@
-export { default } from "./ContributorDetailLayout";
+export { default as ContributorDetailLayout } from "./ContributorDetailLayout";
+export { default as ContributorItemDetailLayout } from "./ContributorItemDetailLayout";
+export { default as ContributorCollectionDetailLayout } from "./ContributorCollectionDetailLayout";
